Crawler is fully working, now comes the finer parts! See crawler_TODO for more details on upcoming changes!
The smb-crawler can be tested:
It needs pysmbc==1.0.9, sqlalchemy>=0.7, whatever provides transaction, and because i didn't comment it out: ftputil
i always do: 
$ rm devdata.db; paster setup-app development.ini
$ python crawler.py smb://user:pass@host/

======================================================
TODO 
for the new, modular crawler engine of Noodle NG:

- Standardized Logging (Really much output in debug,
  only host-based information in info and failures etc in
  warnings and up)
- Better Error/Exception handling from database and filesystem
  (without loosing traceback information, just in case)
- User-Defined Handler-Functions for special tasks
  (e.g. Special file .noodle in root-directory in which hosts
  can specify files/directories to exclude etc)
- More controllable debug-modes (disable/enable multiprocessing,
  level of output, etc)
- Think about all the Unicode-Stuff again... 
- Port changes in CrawlerSMB to CrawlerFTP
- Local file crawler (shouldn't be much work) -> CrawlerLOCAL

- Reconcile classes... Propably it would be senseful to have
  a SMBHost-Object like ftputil provides and use that for all
  filesystem operations (proxyDownloader in Mind) 


git-svn-id: https://noodle-ng.googlecode.com/svn/branches/tg2.1@158 8aa6f4f4-6411-944d-c798-68ba1b83f409

+++ noodle/model/share.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from sqlalchemy import *
 from sqlalchemy.orm import mapper, relationship, backref
 from sqlalchemy import Table, ForeignKey, Column
-from sqlalchemy.types import Integer, Unicode, BigInteger, DateTime#, Float, Numeric
+from sqlalchemy.types import Integer, Unicode, BigInteger, DateTime, Float
 
 import noodle.model
 from noodle.model import DeclarativeBase, metadata, DBSession
@@ -224,7 +224,7 @@ class Host(DeclarativeBase):
     name = Column(Unicode(256))
     first_seen = Column(DateTime, nullable=False)
     last_update = Column(DateTime, nullable=False)
-    crawl_time = Column(Integer)
+    crawl_time = Column(Float)
     sharesize = Column(BigInteger)
     services = relationship(Service, backref="host")
     statistics = relationship(Statistic, backref="host")
